+++ common/src/main/java/tocraft/craftedcore/mixin/client/ClientPacketListenerMixin.java
@@ -17,6 +17,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import tocraft.craftedcore.event.client.ClientPlayerEvents;
 
+@SuppressWarnings("unused")
 @Environment(EnvType.CLIENT)
 @Mixin(ClientPacketListener.class)
 public abstract class ClientPacketListenerMixin extends ClientCommonPacketListenerImpl {