+++ common/src/main/java/tocraft/craftedcore/mixin/client/MinecraftClientMixin.java
@@ -13,6 +13,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import tocraft.craftedcore.event.client.ClientPlayerEvents;
 
+@SuppressWarnings("unused")
 @Environment(EnvType.CLIENT)
 @Mixin(Minecraft.class)
 public abstract class MinecraftClientMixin {