+++ common/src/main/java/tocraft/craftedcore/registration/PlayerDataRegistry.java
@@ -10,7 +10,7 @@ import java.util.Set;
 
 @SuppressWarnings("unused")
 public class PlayerDataRegistry {
-    private static final Map<String, Map.Entry<Boolean, Boolean>> CraftedTagKeys = new HashMap<String, Map.Entry<Boolean, Boolean>>();
+    private static final Map<String, Map.Entry<Boolean, Boolean>> CraftedTagKeys = new HashMap<>();
 
     /**
      * Should be called once the player joins and for every key