+++ fabric/src/main/java/tocraft/craftedcore/fabric/CraftedCoreFabric.java
@@ -1,16 +1,15 @@
 package tocraft.craftedcore.fabric;
 
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.entity.event.v1.EntitySleepEvents;
 import tocraft.craftedcore.CraftedCore;
-import tocraft.craftedcore.event.common.PlayerEvents;
 
+@SuppressWarnings("unused")
 public class CraftedCoreFabric implements ModInitializer {
 
     @Override
     public void onInitialize() {
-        new CraftedCore().initialize();
+        CraftedCoreFabricEventHandler.initialize();
 
-        EntitySleepEvents.ALLOW_SLEEP_TIME.register((player, sleepingPos, vanillaResult) -> PlayerEvents.ALLOW_SLEEP_TIME.invoke().allowSleepTime(player, sleepingPos, vanillaResult));
+        new CraftedCore().initialize();
     }
 }