+++ fabric/src/main/java/tocraft/craftedcore/fabric/client/CraftedCoreFabricClient.java
@@ -1,17 +1,15 @@
-package tocraft.craftedcore.fabric;
+package tocraft.craftedcore.fabric.client;
 
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 import tocraft.craftedcore.client.CraftedCoreClient;
-import tocraft.craftedcore.event.client.RenderEvents;
 
 @Environment(EnvType.CLIENT)
 public class CraftedCoreFabricClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
-        HudRenderCallback.EVENT.register((graphics, tickDelta) -> RenderEvents.HUD_RENDERING.invoke().render(graphics, tickDelta));
+        CraftedCoreFabricEventHandlerClient.initialize();
 
         new CraftedCoreClient().initialize();
     }