+++ neoforge/src/main/java/tocraft/craftedcore/neoforge/CraftedCoreNeoForgeEventHandler.java
@@ -5,12 +5,17 @@ import net.minecraft.world.InteractionResult;
 import net.neoforged.bus.api.Event;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.neoforge.event.AddReloadListenerEvent;
+import net.neoforged.neoforge.event.RegisterCommandsEvent;
+import net.neoforged.neoforge.event.entity.living.LivingBreatheEvent;
 import net.neoforged.neoforge.event.entity.living.LivingDeathEvent;
 import net.neoforged.neoforge.event.entity.player.SleepingTimeCheckEvent;
+import net.neoforged.neoforge.event.level.LevelEvent;
 import net.neoforged.neoforge.event.level.SleepFinishedTimeEvent;
 import tocraft.craftedcore.data.SynchronizedJsonReloadListener;
+import tocraft.craftedcore.event.common.CommandEvents;
 import tocraft.craftedcore.event.common.EntityEvents;
 import tocraft.craftedcore.event.common.PlayerEvents;
+import tocraft.craftedcore.event.common.ServerLevelEvents;
 import tocraft.craftedcore.registration.SynchronizedReloadListenerRegistry;
 
 @SuppressWarnings("unused")
@@ -32,7 +37,7 @@ public class CraftedCoreNeoForgeEventHandler {
 
     @SubscribeEvent
     public void allowSleepTime(SleepingTimeCheckEvent event) {
-        InteractionResult result = PlayerEvents.ALLOW_SLEEP_TIME.invoke().allowSleepTime(event.getEntity(),  event.getSleepingLocation().isPresent() ? event.getSleepingLocation().get() : null, event.getResult() != Event.Result.DENY);
+        InteractionResult result = PlayerEvents.ALLOW_SLEEP_TIME.invoke().allowSleepTime(event.getEntity(), event.getSleepingLocation().isPresent() ? event.getSleepingLocation().get() : null, event.getResult() != Event.Result.DENY);
         if (result == InteractionResult.FAIL) {
             event.setResult(Event.Result.DENY);
         }
@@ -45,5 +50,29 @@ public class CraftedCoreNeoForgeEventHandler {
     public void sleepFinishedTime(SleepFinishedTimeEvent event) {
         long newTimeIn = PlayerEvents.SLEEP_FINISHED_TIME.invoke().setTimeAddition((ServerLevel) event.getLevel(), event.getNewTime());
         event.setTimeAddition(newTimeIn);
+    }
+
+    @SubscribeEvent
+    public void registerCommands(RegisterCommandsEvent event) {
+        CommandEvents.REGISTRATION.invoke().register(event.getDispatcher(), event.getBuildContext(), event.getCommandSelection());
+    }
+
+    @SubscribeEvent
+    public void serverLoad(LevelEvent.Load event) {
+        if (event.getLevel() instanceof ServerLevel serverLevel) {
+            ServerLevelEvents.LEVEL_LOAD.invoke().call(serverLevel);
+        }
+    }
+
+    @SubscribeEvent
+    public void serverUnload(LevelEvent.Unload event) {
+        if (event.getLevel() instanceof ServerLevel serverLevel) {
+            ServerLevelEvents.LEVEL_UNLOAD.invoke().call(serverLevel);
+        }
+    }
+
+    @SubscribeEvent
+    public void livingBreathe(LivingBreatheEvent event) {
+        event.setCanBreathe(EntityEvents.LIVING_BREATHE.invoke().breathe(event.getEntity(), event.canBreathe()));
     }
 }