+++ neoforge/src/main/java/tocraft/craftedcore/neoforge/client/CraftedCoreNeoForgeEventHandlerClient.java
@@ -7,6 +7,8 @@ import net.neoforged.api.distmarker.OnlyIn;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.neoforge.client.event.RenderGuiEvent;
 import net.neoforged.neoforge.client.event.RenderGuiOverlayEvent;
+import net.neoforged.neoforge.event.TickEvent;
+import tocraft.craftedcore.event.client.ClientTickEvents;
 import tocraft.craftedcore.event.client.RenderEvents;
 
 @SuppressWarnings("unused")
@@ -15,6 +17,15 @@ public class CraftedCoreNeoForgeEventHandlerClient {
     @SubscribeEvent
     public void event(RenderGuiEvent.Post event) {
         RenderEvents.HUD_RENDERING.invoke().render(event.getGuiGraphics(), event.getPartialTick());
+    }
+
+    @SubscribeEvent
+    public void event(TickEvent.ClientTickEvent event) {
+        if (event.phase == TickEvent.Phase.START) {
+            ClientTickEvents.CLIENT_PRE.invoke().tick(Minecraft.getInstance());
+        } else if (event.phase == TickEvent.Phase.END) {
+            ClientTickEvents.CLIENT_POST.invoke().tick(Minecraft.getInstance());
+        }
     }
 
     @SubscribeEvent