+++ neoforge/src/main/java/tocraft/craftedcore/platform/neoforge/PlatformDataImpl.java
@@ -1,0 +1,42 @@
+package tocraft.craftedcore.platform.neoforge;
+
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.fml.ModList;
+import net.neoforged.fml.loading.FMLLoader;
+import net.neoforged.fml.loading.FMLPaths;
+import org.jetbrains.annotations.Nullable;
+
+import java.lang.module.ModuleDescriptor.Version;
+import java.nio.file.Path;
+
+@SuppressWarnings("unused")
+public final class PlatformDataImpl {
+    public static boolean isModLoaded(String modid) {
+        return ModList.get().isLoaded(modid);
+    }
+
+    @Nullable
+    public static Version getModVersion(String modid) {
+        return ModList.get().getModContainerById(modid).map(modContainer -> Version.parse(modContainer.getModInfo().getVersion().toString())).orElse(null);
+    }
+
+    public static boolean isDevEnv() {
+        return !FMLLoader.isProduction();
+    }
+
+    public static Dist getEnv() {
+        return FMLLoader.getDist();
+    }
+
+    public static Path getConfigPath() {
+        return FMLPaths.CONFIGDIR.get();
+    }
+
+    public static ModLoader getModLoaderId() {
+        return ModLoader.NEOFORGE;
+    }
+
+    public enum ModLoader {
+        FABRIC, FORGE, NEOFORGE, OTHER
+    }
+}