+++ neoforge/src/main/java/tocraft/craftedcore/registration/neoforge/SynchronizedReloadListenerRegistryImpl.java
@@ -1,0 +1,11 @@
+package tocraft.craftedcore.registration.neoforge;
+
+import net.minecraft.resources.ResourceLocation;
+import tocraft.craftedcore.data.SynchronizedJsonReloadListener;
+
+@SuppressWarnings("unused")
+public class SynchronizedReloadListenerRegistryImpl {
+    @SuppressWarnings("EmptyMethod")
+    public static void onRegister(SynchronizedJsonReloadListener reloadListener, ResourceLocation id) {
+    }
+}